% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhmbasket.predict-package.R
\docType{package}
\name{bhmbasket.predict-package}
\alias{bhmbasket.predict-package}
\alias{bhmbasket.predict}
\title{The 'bhmbasket.predict' package.}
\description{
This package implements methods to dynamically predict response status
of individuals in a basket trial with delayed binary response endpoint
(e.g. tumor response) at arbitrary time points.
}
\references{
Stan Development Team (2021). RStan: the R interface to Stan. R package version 2.21.3. https://mc-stan.org
}
