% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check a visits data set for correct format}
\usage{
check_data(data, model)
}
\arguments{
\item{data}{data.frame to check}

\item{model}{\link{srpmodel} object used to fit data}
}
\value{
data.frame, same as input but all censoring events after terminal
states are removed.
}
\description{
Raises specific errors when encountering issues in the data.
}
\examples{
tbl <- data.frame(group_id = "A", subject_id = "A1", t = 0, state = "stable")
mdl <- create_srpmodel(A = define_srp_prior())
check_data(tbl, mdl)
}
