% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_model.R
\name{create_srp_model}
\alias{create_srp_model}
\title{todo}
\usage{
create_srp_model(
  group_id,
  logodds_mean,
  logodds_sd,
  median_time_to_next_event_mean,
  median_time_to_next_event_sd,
  visit_spacing,
  logodds_min = rep(logodds(0.001), length(group_id)),
  logodds_max = rep(logodds(0.999), length(group_id)),
  shape_min = matrix(0.99, nrow = length(group_id), ncol = 3),
  shape_max = matrix(1.01, nrow = length(group_id), ncol = 3)
)
}
\description{
todo
}
