% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_samples.R
\name{draw_samples}
\alias{draw_samples}
\title{Draw samples from a model}
\usage{
draw_samples(model, ...)
}
\arguments{
\item{model}{the model to simulate from}
}
\description{
simulate exact response times from the specified model,
optionally conditioning on additional data.
}
