% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_visit_data.R
\name{generate_visit_data}
\alias{generate_visit_data}
\title{Generate visit data}
\usage{
generate_visit_data(
  group_id,
  n,
  response_rate,
  delay = numeric(length(group_id)),
  recruitment_rate,
  visit_spacing,
  max_duration,
  responder_weibull_scale,
  responder_weibull_shape,
  nonresponder_weibull_scale,
  nonresponder_weibull_shape,
  seed = NULL
)
}
\arguments{
\item{group_id}{ids of arms}

\item{n}{sample size of arms}

\item{response_rate}{response rates of arms}

\item{delay}{delay (in months) of recruitment per arm}

\item{recruitment_rate}{recruitment rates (per month) of arms}

\item{visit_spacing}{distance between visits (in months)}

\item{max_duration}{maximal overall duration of all arms}

\item{responder_weibull_scale}{parameter of Weibull distribution for TTE responders}

\item{responder_weibull_shape}{parameter of Weibull distribution for TTE responders}

\item{nonresponder_weibull_scale}{parameter of Weibull distribution for TTE non-responders (progression)}

\item{nonresponder_weibull_shape}{parameter of Weibull distribution for TTE non-responders (progression)}

\item{seed}{random seed to use, NULL uses no seed}
}
\value{
A data frame with sampled data in columns "group_id", "subject_id",
"t" (visit time point in months since start of trial), "status" (S(table), R(esponse), P(rogression)),
"eof" (end of follow-up, is the visit the last visit of the individual?)
}
\description{
Generate example data from mixture cure rate model with Weibull distributed
time-to-response and time-to-progression.
}
