% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{impute}
\alias{impute}
\title{Impute data from predictive distribution}
\usage{
impute(model, data, nsim, n_per_group, now, seed, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{data}{the (multi-state) data frame to impute further trajectories for.}

\item{nsim}{number of samples to draw}

\item{n_per_group}{the number of individuals per group to be recruited.}

\item{now}{exact time point relative to start of the trial}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}
}
\value{
a data frame with imputed version of the input data.
}
\description{
If no parameter sample is provided, sample from posterior predictive
}
