% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{impute_predictive}
\alias{impute_predictive}
\title{Impute data from predictive distribution}
\usage{
impute_predictive(
  model,
  data,
  sample,
  nsim,
  nsim_parameters,
  warmup_parameters,
  seed,
  ...
)
}
\arguments{
\item{model}{a model of class "Model"}

\item{data}{the (multi-state) data frame to impute further trajectories for.}

\item{sample}{a stanfit object containing samples. These parameter samples
represent the parameter distribution over which the predictive distribution
averages. Technically, the parameters are resampled with replacement from
this sample to match the desired number of imputations.}

\item{nsim}{number of samples to draw}

\item{nsim_parameters}{integer, number of parameter samples}

\item{warmup_parameters}{integer, number of warmup samples for the rstan sampler before retaining samples of the parameters.}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}
}
\value{
a data frame with imputed version of the input data.
}
\description{
If no parameter sample is provided, sample from posterior predictive
}
