% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndependentMixtureCureRateModel.R
\name{new_IndependentMixtureCureRateModel}
\alias{new_IndependentMixtureCureRateModel}
\title{Create an instance of the mixture cure-rate model for binary tumor response}
\usage{
new_IndependentMixtureCureRateModel(
  group_id,
  logodds_mean,
  logodds_sd,
  logodds_min = rep(logodds(0.001), length(group_id)),
  logodds_max = rep(logodds(0.999), length(group_id)),
  log_shape_mean,
  log_shape_sd,
  median_time_to_response_mean,
  median_time_to_response_sd,
  max_time_to_response,
  visit_spacing,
  monthly_rate_mean,
  monthly_rate_sd
)
}
\arguments{
\item{group_id}{character vector of group ids}

\item{logodds_mean}{numeric vector with the means of the normal prior on the logodds of a response per group}

\item{logodds_sd}{numeric vector with the standard deviations of the normal prior on the logodds of a response per group}

\item{logodds_min}{numeric vector with the minimal logodds per group}

\item{logodds_max}{numeric vector with the maximal logodds per group}

\item{log_shape_mean}{numeric vector with the means of the normal prior on the log-shape parameter of the Weibull distribution for time to response}

\item{log_shape_sd}{numeric vector with the standard deviations of the normal prior on the log-shape parameter of the Weibull distribution for time to response}

\item{median_time_to_response_mean}{numeric vector with the means of the normal prior on the median time to response (truncated at 0)}

\item{median_time_to_response_sd}{numeric vector with the standard deviations of the normal prior on the median for time to response}

\item{max_time_to_response}{numeric vector with the maximal time to response per group}

\item{visit_spacing}{vector of deterministic spacing between future visits (in months)}

\item{monthly_rate_mean}{expected monthly recruitment rate (prior mean)}

\item{monthly_rate_sd}{standard deviation of the expected recruitment rate prior}
}
\value{
An object of class "IndependentMixtureCureRateModel" holding all relevant
prior information.
}
\description{
This model assumes that the response rates across groups are independent.
}
