% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndependentPoissonRecruitmentModel.R
\name{independent_poisson_recruitment_model}
\alias{independent_poisson_recruitment_model}
\alias{draw_samples.IndependentPoissonRecruitmentModel}
\alias{plot.IndependentPoissonRecruitmentModel}
\title{Create an instance of a Poisson recruitment model}
\usage{
independent_poisson_recruitment_model(
  group_id,
  log_monthly_rate_mean,
  log_monthly_rate_sd,
  maximal_recruitment_interval
)

\method{draw_samples}{IndependentPoissonRecruitmentModel}(
  model,
  data = NULL,
  n = NULL,
  now,
  nsim,
  nsim_min = 1000L,
  seed = NULL,
  warmup = 1000L,
  verbose = FALSE,
  show_messages = FALSE,
  refresh = 0,
  return_raw_stan_output = FALSE,
  ...
)

\method{plot}{IndependentPoissonRecruitmentModel}(x, n, nsim = 10000, seed = NULL, ...)
}
\description{
Create an instance of a Poisson recruitment model
}
\section{Functions}{
\itemize{
\item \code{draw_samples.IndependentPoissonRecruitmentModel}: draw samples

\item \code{plot.IndependentPoissonRecruitmentModel}: plot prior assumptions
}}

