% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_model.R
\name{is_valid.srp_model}
\alias{is_valid.srp_model}
\title{To check the validity of input parameters to create_srp}
\usage{
\method{is_valid}{srp_model}(mdl)
}
\arguments{
\item{mdl}{an object of type srp_model}
}
\value{
True statements or corresponding parameter specific errors
}
\description{
To check the validity of input parameters to create_srp
}
