% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstate_to_visit.R
\name{mstate_to_visits}
\alias{mstate_to_visits}
\title{Convert data from multi-state to visit format}
\usage{
mstate_to_visits(model, tbl_mstate, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{tbl_mstate}{a data table in multi-state format. TODO: explain}

\item{...}{further arguments passed to method implementations}
}
\value{
A data frame (tibble) in visit format, where each line corresponds to
individual visits.
}
\description{
This is still experimental since the conversion is not 1:1. The multi-state
representation drops non-changing visits, hence the exact visit data cannot
be recovered form a multi-state representation.
}
\details{
The intermediate visits are reconstructed using the visit spacing
information stored in the model.
}
