% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncomsm-package.R
\docType{package}
\name{oncomsm-package}
\alias{oncomsm-package}
\alias{oncomsm}
\title{The oncomsm package}
\description{
This package implements methods to dynamically predict response
and progression of individuals in early oncology trials using parametric
multi-state models and Bayesian inference.
This allows the dynamic computation of Probability of Success for a wide
The inference is implemented using 'stan'.
}
\references{
Stan Development Team (2021). "RStan: the R interface to Stan".
R package version 2.21.3. https://mc-stan.org
}
