% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{parameter_sample_to_tibble}
\alias{parameter_sample_to_tibble}
\title{Convert stanfit sample to data table}
\usage{
parameter_sample_to_tibble(model, sample, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{sample}{a stanfit object with samples from the respective model.}

\item{...}{further arguments passed to method implementations}
}
\value{
a tibble with the sampled parameters in long format
}
\description{
Convert stanfit sample to data table
}
