% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pfs.R
\name{plot_pfs}
\alias{plot_pfs}
\title{Plot progression-free-survival function}
\usage{
plot_pfs(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  confidence = NULL,
  ...
)
}
\arguments{
\item{model}{an object of class \link{srpmodel} containing prior information}

\item{parameter_sample}{a stanfit object with samples from the respective model.}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{nsim}{integer, number of samples to draw}

\item{warmup}{integer, number of warm-up samples for the MCMC sampler
before retaining samples; see \code{warmup} parameter in \code{\link[rstan:stan]{rstan::stan()}}.}

\item{nuts_control}{list, parameters for NUTS algorithm see \code{control}
argument in\code{\link[rstan:stan]{rstan::stan()}}}

\item{dt_interval}{numeric vector of length two with minimal and maximal
time (relative to individual first visit) to use for plotting}

\item{dt_n_grid}{number of grid points to use when automatically choosing
plotting interval}

\item{dt_expand}{expansion factor for upper plotting limit when using
automatic interval detection}

\item{dt_grid}{numeric vector of time points to use for plotting}

\item{confidence}{numeric in (0, 1) confidence level for point-wise
confidence bands around mean; none plotted if NULL.}

\item{...}{further arguments passed to method implementations}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\description{
\code{plot_pfs()} plots the progression-free-survival function of a model.
}
\examples{
\dontrun{
mdl <- create_srpmodel(A = define_srp_prior())
plot_pfs(mdl)
}
}
\seealso{
\code{\link[=plot_transition_times]{plot_transition_times()}} \code{\link[=plot_response_probability]{plot_response_probability()}}
}
