% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transition_times.R
\name{plot_transition_times}
\alias{plot_transition_times}
\alias{plot_transition_times.model}
\title{Plot the transition times of a model}
\usage{
plot_transition_times(model, parameter_sample, ...)

\method{plot_transition_times}{model}(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  confidence = NULL,
  ...
)
}
\arguments{
\item{model}{an object of class \link{model} containing prior information}

\item{parameter_sample}{a stanfit object with samples from the respective model.}

\item{...}{further arguments passed to method implementations}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{nsim}{integer, number of samples to draw}

\item{warmup}{integer, number of warm-up samples for the MCMC sampler
before retaining samples; see \code{warmup} parameter in \code{\link[rstan:stan]{rstan::stan()}}.}

\item{dt_interval}{numeric vector of length two with minimal and maximal
time (relative to individual first visit) to use for plotting}

\item{dt_n_grid}{number of grid points to use when automatically choosing
plotting interval}

\item{dt_expand}{expansion factor for upper plotting limit when using
automatic interval detection}

\item{dt_grid}{numeric vector of time points to use for plotting}

\item{confidence}{numeric in (0, 1) confidence level for point-wise
confidence bands around mean; none plotted if NULL.}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\description{
\code{plot_transition_times()} plots a the survival functions for the transition
times in a multi-state model.
}
\examples{
mdl <- create_model(A = group_prior())
plot_transition_times(mdl)

}
\seealso{
\code{\link[=plot_pfs]{plot_pfs()}} \code{\link[=plot_response_probability]{plot_response_probability()}}
}
