% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visits.R
\name{plot_visits}
\alias{plot_visits}
\title{Plot a data sets of visits}
\usage{
plot_visits(visit_data, epsilon = 5/30, event_positive = TRUE)
}
\arguments{
\item{visit_data}{data frame as generated by \link{generate_visit_data}}

\item{epsilon}{numeric width (in months) of final status indicator}

\item{event_positive}{logical, is the event of interest positive (e.g. a response to treatment)? If TRUE, the events are coded in green.}
}
\description{
Plot a data sets of visits
}
