% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{sample_pfs_rate}
\alias{sample_pfs_rate}
\title{Sample from the progression-free-survival rate}
\usage{
sample_pfs_rate(model, t, sample, warmup, nsim, seed, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{t}{a vector of time-points at which the PFS rate should be computed}

\item{sample}{a stanfit object with samples from the respective model.}

\item{warmup}{integer, number of warmup samples for the rstan sampler before retaining samples; these are used to
tune the hyperparameters of the MCMC algorithm.}

\item{nsim}{number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}
}
\value{
a data frame with samples of PFS rates at each of the time points
in the vector t.
}
\description{
Progression-free-survival rate at time t (PFS-t rate) is a function of the
parameters of a given multi-state model. Hence any prior or posterior sample
from such a model gives rise to a sample of the corresponding PFS t rate.
}
