% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{sample_posterior}
\alias{sample_posterior}
\alias{sample_posterior.model}
\title{Sample from posterior distribution of a model}
\usage{
sample_posterior(model, data, nsim, seed, ...)

\method{sample_posterior}{model}(
  model,
  data,
  nsim = 2000L,
  seed = NULL,
  warmup = 500L,
  nuts_control = list(),
  pars = attr(model, "parameter_names"),
  ...
)
}
\arguments{
\item{model}{a model of class "Model"}

\item{data}{a data frame with variables "subject_id", "group_id", "t_recruitment", "dt1" and "dt2"
where dt1 is the minimal and dt2 the maximal time to the event in question.
If both are Inf, the individual has not and will never experience the event.
If dt1 < Inf but dt2 == Inf the individual is still at risk.
"t_recruitment", "dt1" and "dt2" can also be missing if individual is not yet recruited.}

\item{nsim}{number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}

\item{warmup}{integer, number of warmup samples for the rstan sampler before retaining samples; these are used to
tune the hyperparameters of the MCMC algorithm.}

\item{nuts_control}{control parameters for NUTS algorithm see \code{\link[rstan:stan]{rstan::stan()}}}

\item{pars}{character vector of parameter names to sample; resorting to model defaults if NULL}
}
\value{
A rstanfit object with posterior samples.
}
\description{
\code{sample_posterior()} draws samples from the
posterior distribution of the specified model given a data set with
visit data
}
\seealso{
\code{\link[=parameter_sample_to_tibble]{parameter_sample_to_tibble()}}
}
