% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_predictive.R
\name{sample_predictive}
\alias{sample_predictive}
\alias{sample_predictive.model}
\title{Sample visits from predictive distribution}
\usage{
sample_predictive(model, n_per_group, sample, nsim, seed, ...)

\method{sample_predictive}{model}(
  model,
  n_per_group,
  sample = NULL,
  nsim = 100L,
  seed = NULL,
  nsim_parameters = 1000L,
  warmup_parameters = 250,
  as_mstate = FALSE,
  nuts_control = list(),
  ...
)
}
\arguments{
\item{model}{an object of class \link{model} containing prior information}

\item{n_per_group}{integer vector with number of individuals per group.}

\item{sample}{a stanfit object with samples from the respective model.}

\item{nsim}{integer, number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}

\item{nsim_parameters}{integer, number of parameter samples}

\item{warmup_parameters}{integer, number of warmup samples for the rstan sampler before retaining samples of the parameters.}

\item{as_mstate}{return data in multi-state forma, see \code{\link[=visits_to_mstate]{visits_to_mstate()}}}

\item{nuts_control}{list, parameters for NUTS algorithm see \code{control}
argument in\code{\link[rstan:stan]{rstan::stan()}}}
}
\value{
a data frame with variables
\verb{subject_id<chr>} (subject identifier),
\verb{group_id<chr>} (group identifier),
\verb{t<dbl>} (time of visit, relative to first visit in study),
\verb{state<chr>} (state recorded at visit)
\verb{iter<int>} (re-sample indicator).
Allowed states are "stable", "response", "progression" (or death),
and "EOF" (end of follow-up).
The EOF state marks the end of an individual's follow-up before the absorbing
state "progression".
}
\description{
\code{sample_predictive()} draws samples from the predictive distribution of a
model given a parameter sample.
}
\examples{
mdl <- create_model(A = group_prior())
sample_predictive(mdl, 1L, nsim = 1L, seed = 38L)

}
\seealso{
\code{\link[=sample_prior]{sample_prior()}} \code{\link[=sample_posterior]{sample_posterior()}} \code{\link[=impute]{impute()}}
}
