% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(
  model,
  n_to_be_recruited,
  nsim,
  data = NULL,
  seed = 42L,
  warmup = 1000L,
  ...
)
}
\arguments{
\item{model}{the model to simulate from}

\item{data}{\strong{optional} data to condition on}
}
\description{
simulate exact response time points from a Weibull mixture cure rate model.
}
