% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_model.R, R/srp_model_plotting.R
\name{srp_model}
\alias{srp_model}
\alias{srp_group_prior}
\alias{create_srp_model}
\alias{format.srp_model}
\alias{visits_to_mstate.srp_model}
\alias{compute_pfs.srp_model}
\alias{plot_mstate.srp_model}
\alias{plot_transition_times.srp_model}
\alias{plot_response_probability.srp_model}
\alias{plot_pfs.srp_model}
\alias{plot.srp_model}
\title{Combine SRP group priors to model object}
\usage{
srp_group_prior(
  p_mean = 0.5,
  p_n = 3,
  p_eta = 0,
  p_min = 0,
  p_max = 1,
  median_t_q05 = c(1, 1, 3),
  median_t_q95 = c(12, 12, 24),
  shape_q05 = c(0.99, 0.99, 0.99),
  shape_q95 = c(1.01, 1.01, 1.01),
  visit_spacing = 1,
  recruitment_rate = 1
)

create_srp_model(..., maximal_time = 10 * 12)

\method{format}{srp_model}(x, ...)

\method{visits_to_mstate}{srp_model}(
  tbl_visits,
  model,
  now = max(tbl_visits$t),
  eof_indicator = "EOF"
)

\method{compute_pfs}{srp_model}(
  model,
  t,
  parameter_sample = NULL,
  warmup = 500L,
  nsim = 1000L,
  seed = NULL,
  ...
)

\method{plot_mstate}{srp_model}(
  data,
  model,
  now = max(tbl_mstate$t_max),
  relative_to_sot = TRUE,
  ...
)

\method{plot_transition_times}{srp_model}(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  confidence = NULL,
  ...
)

\method{plot_response_probability}{srp_model}(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  ...
)

\method{plot_pfs}{srp_model}(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  confidence = NULL,
  ...
)

\method{plot}{srp_model}(
  x,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  confidence = NULL,
  ...
)
}
\arguments{
\item{p_mean}{numeric, mean of the beta prior for the response probability}

\item{p_n}{numeric, beta prior equivalent sample size (a + b)}

\item{p_eta}{numeric, robustification parameter for beta prior; actual
prior is (1 - eta) beta + eta; i.e., eta is the non-informative weight.}

\item{p_min}{numeric, minimal response probability}

\item{p_max}{numeric, maximal response probability}

\item{median_t_q05}{numeric of length three,
5\% quantiles of the log-normal distributions for the
median time-to-next-event for the three transitions s->r, s->p, r->p.}

\item{median_t_q95}{numeric of length three,
95\% quantiles of the log-normal distributions for the
median time-to-next-event for the three transitions s->r, s->p, r->p.}

\item{shape_q05}{numeric of length three,
5\% quantiles of the log-normal distributions for the shapes of the
time-to-next-event distributions for the three transitions s->r, s->p, r->p.}

\item{shape_q95}{numeric of length three,
95\% quantiles of the log-normal distributions for the shapes of the
time-to-next-event distributions for the three transitions s->r, s->p, r->p.}

\item{visit_spacing}{numeric, fixed duration between visits}

\item{recruitment_rate}{numeric, constant recruitment rate}

\item{...}{further arguments passed to method implementations}

\item{maximal_time}{the maximal overall runtime of the trial as measured from
the first visit of any group. No visits past this point are sampled.}

\item{x}{the model to plot}

\item{tbl_visits}{visit data in long format}

\item{model}{a multi-state model object}

\item{now}{time point since start of trial (might be later than last
recorded visit)}

\item{eof_indicator}{state name indicating (exactly observed) end of
follow up.}

\item{t}{a vector of time-points at which the PFS rate should be computed}

\item{parameter_sample}{a stanfit object with samples from the respective model.}

\item{warmup}{integer, number of warmup samples for the rstan sampler before retaining samples; these are used to
tune the hyperparameters of the MCMC algorithm.}

\item{nsim}{number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{data}{a data table with multi-state data}

\item{relative_to_sot}{Boolean, should the timeline be relative to the start
of trial or the start of treatment for each individual}

\item{nuts_control}{control parameters for NUTS algorithm see \code{\link[rstan:stan]{rstan::stan()}}}

\item{dt_interval}{numeric vector of length two with minimal and maximal
time (relative to individual first visit) to use for plotting}

\item{dt_n_grid}{number of grid points to use when automatically choosing
plotting interval}

\item{dt_expand}{expansion factor for upper plotting limit when using
automatic interval detection}

\item{dt_grid}{numeric vector of time points to use for plotting}

\item{confidence}{numeric in (0, 1) confidence level for point-wise
confidence bands around mean; none plotted if NULL.}
}
\value{
An object of type srp_group_prior holding all prior information in a
list-like structure; visit_spacing and recruitment_rate are accessible as
attributes.

an object of class srp_model; a named list of the group_ids,
the maximal time, the visit spacing, the recruitment rate, and a list of
prior parameters for the response probability, the median transition times,
and the shape of the transition time distribution (Weibull)
}
\description{
This function takes one or more prior-specifications for an SRP multi-state
model and combines them into a joint model.
Groups are still treated as independent.

This function takes one or more prior-specifications for an SRP multi-state
model and combines them into a joint model.
Groups are still treated as independent.
}
