% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srpmodel.R
\name{srpmodel}
\alias{srpmodel}
\alias{srp-model}
\alias{define_srp_prior}
\alias{create_srpmodel}
\title{A stable-response-progression model}
\usage{
define_srp_prior(
  p_mean = 0.5,
  p_n = 3,
  p_eta = 0,
  p_min = 0,
  p_max = 1,
  median_t_q05 = c(1, 1, 3),
  median_t_q95 = c(12, 12, 24),
  shape_q05 = rep(0.99, 3),
  shape_q95 = rep(1.01, 3),
  visit_spacing = 1,
  recruitment_rate = 1
)

create_srpmodel(..., maximal_time = 10 * 12)
}
\arguments{
\item{p_mean}{numeric, mean of the beta prior for the response probability}

\item{p_n}{numeric, beta prior equivalent sample size (a + b)}

\item{p_eta}{numeric, robustification parameter for beta prior; actual
prior is (1 - eta) beta + eta; i.e., eta is the non-informative weight.}

\item{p_min}{numeric, minimal response probability}

\item{p_max}{numeric, maximal response probability}

\item{median_t_q05}{numeric of length three,
5\% quantiles of the log-normal distributions for the
median time-to-next-event for the three transitions s->r, s->p, r->p.}

\item{median_t_q95}{numeric of length three,
95\% quantiles of the log-normal distributions for the
median time-to-next-event for the three transitions s->r, s->p, r->p.}

\item{shape_q05}{numeric of length three,
5\% quantiles of the log-normal distributions for the shapes of the
time-to-next-event distributions for the three transitions s->r, s->p, r->p.}

\item{shape_q95}{numeric of length three,
95\% quantiles of the log-normal distributions for the shapes of the
time-to-next-event distributions for the three transitions s->r, s->p, r->p.}

\item{visit_spacing}{numeric, fixed duration between visits}

\item{recruitment_rate}{numeric, constant recruitment rate}

\item{...}{named \code{srp_prior} objects; the argument names serve as
group labels}

\item{maximal_time}{the maximal overall runtime of the trial as measured from
the first visit of any group. No visits past this point are sampled.}
}
\description{
\code{create_model()} takes one or more prior-specifications for an
SRP multi-state model and combines them into a single model object.
Groups are still treated as independent.
}
\details{
\code{define_srp_prior()} specifies a prior distribution for a
three state model (stable, response, progression) for
a single group.
}
\examples{
# a model with prior 25\% response rate and variance equivalent to
# 10 data points (i.e. a Beta(2.5, 7.5) distribution).
grp <- define_srp_prior(p_mean = 0.25, p_n = 10)

# a model with two groups and different priors on the respective response
# probabilities
create_srpmodel(
  A = define_srp_prior(),
  B = define_srp_prior(p_mean = 0.33, p_n = 10)
)

}
