% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_example_data.R
\docType{data}
\name{tbl_example_data}
\alias{tbl_example_data}
\title{Example data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 471 rows and 6 columns.
}
\usage{
tbl_example_data
}
\description{
Example data
}
\keyword{datasets}
