% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_interim_example.R
\docType{data}
\name{tbl_interim_example}
\alias{tbl_interim_example}
\title{A Interim-Analysis Example Data Set}
\format{
\subsection{\code{tbl_interim_example}}{

A data frame with 45 rows and 7 columns:
\describe{
\item{subject_id}{subject identifier}
\item{group_id}{group identifier}
\item{from}{starting state of transition}
\item{to}{end state of transition, missing if censored}
\item{t_min}{lower boundary on jump time (in months)}
\item{t_max}{upper boundary on jump time (in months)}
\item{t_sot}{recrutiment time since start of study (in months)}
}
}
}
\usage{
tbl_interim_example
}
\description{
This data set consists of data from a three-arm study at time point
10 months after start of recruitment.
Jump times of this multistate data set are properly censored.
}
\details{
The data was generated with a recruitment rate of 3/per month (overall) and
aiming at 20 individuals per group.
}
\keyword{datasets}
