% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThisModelNeedsAGoodName.R
\name{this_function_needs_a_good_name}
\alias{this_function_needs_a_good_name}
\title{Fully Exchangeable Model (FEM)}
\usage{
this_function_needs_a_good_name(
  logor_loc,
  logor_scale,
  alpha_loc,
  alpha_scale,
  sigma_loc,
  sigma_scale
)
}
\arguments{
\item{mu}{mean of Gaussian prior for latent heterogeneity}

\item{tau}{standard deviation of Gaussian prior for latent heterogeneity}
}
\value{
An object of class FEM with prior information.
}
\description{
todo
}
\examples{
this_function_needs_a_good_name(0, 0.05, 2, .1, 3, 1)

}
