% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits_to_tte.R
\name{visits_to_tte}
\alias{visits_to_tte}
\title{Convert longitudinal visit data to time-to-event data}
\usage{
visits_to_tte(tbl_visits, cutoff = Inf)
}
\arguments{
\item{tbl_visits}{visit data in long format; columns "group_id" (character),
"subject_id" (character), "t" (numeric, visit time point in months since start of trial),
"status" (character, S/P/R), and "eof"
(logical, end of follow-up, is the visit the last visit of the individual?)}

\item{cutoff}{maximal visit times to consider - allows retrospective analyses.}
}
\value{
A data frame with one row per subject in tbl_visits. columns:
"group_id", "subject_id", "t_recruitment" (recruitment time since start of trial),
"dt1" (minimal time since recruitment to response), "dt2" (maximal time since recruitment to response)
}
\description{
Convert longitudinal visit data to time-to-event data
}
