% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{compute_pfs}
\alias{compute_pfs}
\title{Sample from the progression-free-survival rate}
\usage{
compute_pfs(model, t, parameter_sample, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{t}{a vector of time-points at which the PFS rate should be computed}

\item{parameter_sample}{a stanfit object with samples from the respective model.}

\item{...}{further arguments passed to method implementations}
}
\value{
a data frame with samples of PFS rates at each of the time points
in the vector t.
}
\description{
Progression-free-survival rate at time t (PFS-t rate) is a function of the
parameters of a given multi-state model. Hence any prior or posterior sample
from such a model gives rise to a sample of the corresponding PFS t rate.
}
