% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_samples.R
\name{draw_samples}
\alias{draw_samples}
\title{Draw samples from a model}
\usage{
draw_samples(model, data, n, ...)
}
\arguments{
\item{model}{the model to simulate from}

\item{data}{NULL (then n must be given) or data frame with columns "group_id", "subject_id",
"t_recruitment" (recruitment time since start of trial),
"dt1" (minimal time since recruitment to response, Inf if non-responder),
"dt2" (maximal time since recruitment to response, Inf if non-responder, NA if censored)}

\item{n}{NULL (then data must be provided) or an integer vector with the
per-arm sample sizes for each group_id as given in "model"}
}
\description{
simulate exact response times from the specified model,
optionally conditioning on additional data.
}
