% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_visit_data.R
\name{generate_visit_data}
\alias{generate_visit_data}
\title{Generate visit data}
\usage{
generate_visit_data(
  group_id,
  n,
  event_rate,
  delay = numeric(length(group_id)),
  recruitment_rate,
  visit_spacing,
  max_duration,
  event_weibull_scale,
  event_weibull_shape,
  nonevent_weibull_scale,
  nonevent_weibull_shape,
  seed = NULL
)
}
\arguments{
\item{group_id}{ids of arms}

\item{n}{sample size of arms}

\item{event_rate}{event rates per arm}

\item{delay}{delay (in months) of recruitment per arm}

\item{recruitment_rate}{recruitment rates (per month) of arms}

\item{visit_spacing}{distance between visits (in months)}

\item{max_duration}{maximal overall duration of all arms}

\item{event_weibull_scale}{parameter of Weibull distribution for TTE}

\item{event_weibull_shape}{parameter of Weibull distribution for TTE}

\item{nonevent_weibull_scale}{parameter of Weibull distribution for TTE non-event}

\item{nonevent_weibull_shape}{parameter of Weibull distribution for TTE non-event}

\item{seed}{random seed to use, NULL uses no seed}
}
\value{
A data frame with sampled data in columns "group_id", "subject_id",
"t" (visit time point in months since start of trial), "status" (S(table), E(vent), N(on-event)),
"eof" (end of follow-up, is the visit the last visit of the individual?)
}
\description{
Generate example data from mixture cure rate model with Weibull distributed
time-to-first-event.
It is assumed that there is a competing event that blocks the event of interest
(i.e., an individual can no longer experience the event of interest).
}
