% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{generate_visit_data}
\alias{generate_visit_data}
\title{Generate Visit data from a multi-state model}
\usage{
generate_visit_data(model, n_per_group, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{n_per_group}{integer vector with number of individuals per group.}
}
\description{
Generate Visit data from a multi-state model
}
