% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\alias{impute.model}
\title{Impute visits from predictive distribution}
\usage{
impute(model, data, nsim, n_per_group, now, seed, ...)

\method{impute}{model}(
  model,
  data,
  nsim,
  n_per_group = NULL,
  now = NULL,
  seed = NULL,
  recruitment_rates = model$recruitment_rate,
  sample = NULL,
  nsim_parameters = 1000L,
  warmup_parameters = 250L,
  nuts_control = list(),
  ...
)
}
\arguments{
\item{model}{an object of class \link{model} containing prior information}

\item{data}{a data frame with variables
\verb{subject_id<chr>} (subject identifier),
\verb{group_id<chr>} (group identifier),
\verb{t<dbl>} (time of visit, relative to first visit in study),
\verb{state<chr>} (state recorded at visit).
Allowed states are "stable", "response", "progression" (or death),
and "EOF" (end of follow-up).
The EOF state marks the end of an individual's follow-up before the absorbing
state "progression".}

\item{nsim}{integer, number of samples to draw}

\item{n_per_group}{the number of individuals per group to be recruited.}

\item{now}{exact time point relative to start of the trial}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}

\item{recruitment_rates}{vector of recruitment rates}

\item{sample}{a stanfit object containing samples. These parameter samples
represent the parameter distribution over which the predictive distribution
averages. Technically, the parameters are resampled with replacement from
this sample to match the desired number of imputations.}

\item{nsim_parameters}{integer, number of parameter samples}

\item{warmup_parameters}{integer, number of warmup samples for the rstan sampler before retaining samples of the parameters.}

\item{nuts_control}{list, parameters for NUTS algorithm see \code{control}
argument in\code{\link[rstan:stan]{rstan::stan()}}}
}
\value{
a data frame with variables
\verb{subject_id<chr>} (subject identifier),
\verb{group_id<chr>} (group identifier),
\verb{t<dbl>} (time of visit, relative to first visit in study),
\verb{state<chr>} (state recorded at visit)
\verb{iter<int>} (re-sample indicator).
Allowed states are "stable", "response", "progression" (or death),
and "EOF" (end of follow-up).
The EOF state marks the end of an individual's follow-up before the absorbing
state "progression". The input data is copied for each re-sample and missing
data is sampled from the predictive distribution.
}
\description{
\code{impute()} samples visits for individuals in \code{data} and potentially missing
indivduals up to a maximum of \code{n_per_group} from the posterior
predictive distribution of the given model.
}
\examples{
mdl <- create_model(A = group_prior())
tbl <- tibble::tibble(
  subject_id = c("A1", "A1"),
  group_id = c("A", "A"),
  t = c(0, 1.5),
  state = c("stable", "stable")
)
impute(mdl, tbl, 1L, seed = 38L)

}
\seealso{
\code{\link[=sample_prior]{sample_prior()}} \code{\link[=sample_posterior]{sample_posterior()}} \code{\link[=sample_predictive]{sample_predictive()}}
}
