% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{impute_posterior_predictive}
\alias{impute_posterior_predictive}
\title{Sample from posterior predictive distribution}
\usage{
impute_posterior_predictive(
  model,
  conditional_on,
  parameters,
  nsim = 1000L,
  ...
)
}
\arguments{
\item{model}{the model to sample from}

\item{conditional_on}{data frame with columns "group_id", "subject_id",
"t_recruitment" (recruitment time since start of trial),
"dt1" (minimal time since recruitment to event, Inf if definite no event),
"dt2" (maximal time since recruitment to event, Inf if definite no event or censored)}

\item{parameters}{list of model parameters}

\item{nsim}{number of samples to draw}
}
\value{
...
}
\description{
Sample outcomes from model parameters conditional on a data set (posterior).
}
