% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{impute_posterior_predictive}
\alias{impute_posterior_predictive}
\alias{impute_posterior_predictive.Model}
\title{Sample data from posterior-predictive distribution}
\usage{
impute_posterior_predictive(
  model,
  data,
  now,
  nsim,
  nsim_parameters,
  warmup_parameters,
  seed,
  ...
)

\method{impute_posterior_predictive}{Model}(
  model,
  data,
  now = NULL,
  nsim = 1000L,
  nsim_parameters = 1000L,
  warmup_parameters = 250L,
  seed = NULL,
  ...
)
}
\arguments{
\item{model}{a model of class "Model"}

\item{data}{a data frame with variables "subject_id", "group_id", "t_recruitment", "dt1" and "dt2"
where dt1 is the minimal and dt2 the maximal time to the event in question.
If both are Inf, the individual has not and will never experience the event.
If dt1 < Inf but dt2 == Inf the individual is still at risk.
"t_recruitment", "dt1" and "dt2" can also be missing if individual is not yet recruited.}

\item{now}{numeric, current time in months since start of trial (might be later than last recorded event);
if NULL, last recorded event time is used instead}

\item{nsim}{number of samples to draw}

\item{nsim_parameters}{integer, number of parameter samples}

\item{warmup_parameters}{integer, number of warmup samples for the rstan sampler before retaining samples of the parameters.}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}
}
\value{
data frame with variables "subject_id", "group_id", "t_recruitment", "dt1", "dt2", "dt_eof"
where dt1 is the minimal and dt2 the maximal time to the event in question.
}
\description{
Sample time-to-first event data from given model posterior-predictive
distribution given data.
}
