% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndependentMixtureCureRateModel.R
\name{independent_mixture_cure_rate_model}
\alias{independent_mixture_cure_rate_model}
\alias{draw_samples.IndependentMixtureCureRateModel}
\alias{plot.IndependentMixtureCureRateModel}
\title{Create an instance of the mixture cure-rate model for binary tumor response}
\usage{
independent_mixture_cure_rate_model(
  group_id,
  logodds_mean,
  logodds_sd,
  logodds_min = rep(-Inf, length(group_id)),
  logodds_max = rep(Inf, length(group_id)),
  shape_mean,
  shape_sd,
  scale_mean,
  scale_sd,
  visit_spacing
)

\method{draw_samples}{IndependentMixtureCureRateModel}(
  model,
  data = NULL,
  n = NULL,
  nsim,
  seed = NULL,
  warmup = 1000L,
  verbose = FALSE,
  show_messages = FALSE,
  refresh = 0,
  return_raw_stan_output = FALSE,
  ...
)

\method{plot}{IndependentMixtureCureRateModel}(
  x,
  n = rep(1L, length(attr(model, "group_id"))),
  nsim = 10000,
  seed = NULL,
  ...
)
}
\arguments{
\item{group_id}{character vector of group ids}

\item{logodds_mean}{numeric vector with the means of the normal prior on the logodds of a response per group}

\item{logodds_sd}{numeric vector with the standard deviations of the normal prior on the logodds of a response per group}

\item{logodds_min}{numeric vector with the minimal logodds per group}

\item{logodds_max}{numeric vector with the maximal logodds per group}

\item{shape_mean}{numeric vector with the means of the normal prior on the shape parameter of the Weibull distribution for time to response}

\item{shape_sd}{numeric vector with the standard deviations of the normal prior on the shape parameter of the Weibull distribution for time to response}

\item{scale_mean}{numeric vector with the means of the normal prior on the scale parameter of the Weibull distribution for time to response}

\item{scale_sd}{numeric vector with the standard deviations of the normal prior on the scale parameter of the Weibull distribution for time to response}

\item{visit_spacing}{vector of deterministic spacing between future visits (in months)}
}
\value{
An object of class "IndependentMixtureCureRateModel" holding all relevant
prior information.
}
\description{
This model assumes that the response rates across groups are independent.
}
\section{Functions}{
\itemize{
\item \code{draw_samples.IndependentMixtureCureRateModel}: draw samples

\item \code{plot.IndependentMixtureCureRateModel}: plot prior assumptions
}}

