% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndependentMixtureCureRateModel.R
\name{IndependentMixtureCureRateModel}
\alias{IndependentMixtureCureRateModel}
\alias{plot.IndependentMixtureCureRateModel}
\title{Create an instance of the mixture cure-rate model for binary event endpoint}
\usage{
IndependentMixtureCureRateModel(
  group_id,
  logodds_mean,
  logodds_sd,
  logodds_min = rep(logodds(0.001), length(group_id)),
  logodds_max = rep(logodds(0.999), length(group_id)),
  shape,
  median_time_to_event_mean,
  median_time_to_event_sd,
  max_time_to_event,
  visit_spacing,
  monthly_rate_mean,
  monthly_rate_sd
)

\method{plot}{IndependentMixtureCureRateModel}(
  x,
  sample = NULL,
  data = NULL,
  warmup = 250L,
  nsim = 10000,
  seed = NULL,
  ...
)
}
\arguments{
\item{group_id}{character vector of group ids}

\item{logodds_mean}{numeric vector with the means of the normal prior on the logodds of a response per group}

\item{logodds_sd}{numeric vector with the standard deviations of the normal prior on the logodds of a response per group}

\item{logodds_min}{numeric vector with the minimal logodds per group}

\item{logodds_max}{numeric vector with the maximal logodds per group}

\item{shape}{numeric vector with the shape parameters of the Weibull
distribution for time to event. This is treated as fixed and no inference
on the shape is conducted since it is close to unidentifiable in most
cases.}

\item{median_time_to_event_mean}{numeric vector with the means of the normal prior on the median time to event (truncated at 0)}

\item{median_time_to_event_sd}{numeric vector with the standard deviations of the normal prior on the median for time to event}

\item{max_time_to_event}{numeric vector with the maximal time to event per group}

\item{visit_spacing}{vector of deterministic spacing between future visits (in months)}

\item{monthly_rate_mean}{expected monthly recruitment rate (prior mean)}

\item{monthly_rate_sd}{standard deviation of the expected recruitment rate prior}

\item{x}{model to plot}

\item{sample}{existing parameter sample to plot}

\item{data}{a data frame with variables "subject_id", "group_id", "t_recruitment", "dt1" and "dt2"
where dt1 is the minimal and dt2 the maximal time to the event in question.
If both are Inf, the individual has not and will never experience the event.
If dt1 < Inf but dt2 == Inf the individual is still at risk.
"t_recruitment", "dt1" and "dt2" can also be missing if individual is not yet recruited.}

\item{warmup}{integer, number of warmup samples for the rstan sampler before retaining samples; these are used to
tune the hyperparameters of the MCMC algorithm.}

\item{nsim}{number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}
}
\value{
An object of class "IndependentMixtureCureRateModel" holding all relevant
prior information.

(invisibly) a ggplot2 plot object (composed via patchwork)
}
\description{
This model assumes that the response rates across groups are independent.

visualize a model by sampling from its prior(predictive) distribution.
Alternatively, a parameter sample (e.g. a posterior) can be plotted directly.
}
