% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R, R/srp_model.R, R/srp_model_plotting.R
\name{print.model}
\alias{print.model}
\alias{sample_predictive.model}
\alias{impute.model}
\alias{model}
\alias{group_prior}
\alias{create_model}
\alias{format.model}
\alias{visits_to_mstate.model}
\alias{compute_pfs.model}
\alias{plot_mstate.model}
\alias{plot_transition_times.model}
\alias{plot_response_probability.model}
\alias{plot_pfs.model}
\alias{plot.model}
\title{Combine SRP group priors to model object}
\usage{
\method{print}{model}(x, ...)

\method{sample_predictive}{model}(
  model,
  n_per_group,
  sample = NULL,
  nsim = 100L,
  seed = NULL,
  nsim_parameters = 1000L,
  warmup_parameters = 250,
  as_mstate = FALSE,
  nuts_control = list(),
  ...
)

\method{impute}{model}(
  model,
  data,
  nsim,
  n_per_group = NULL,
  now = NULL,
  seed = NULL,
  recruitment_rates = model$recruitment_rate,
  sample = NULL,
  nsim_parameters = 1000L,
  warmup_parameters = 250L,
  nuts_control = list(),
  ...
)

group_prior(
  p_mean = 0.5,
  p_n = 3,
  p_eta = 0,
  p_min = 0,
  p_max = 1,
  median_t_q05 = c(1, 1, 3),
  median_t_q95 = c(12, 12, 24),
  shape_q05 = rep(0.99, 3),
  shape_q95 = rep(1.01, 3),
  visit_spacing = 1,
  recruitment_rate = 1
)

create_model(..., maximal_time = 10 * 12)

\method{format}{model}(x, ...)

\method{visits_to_mstate}{model}(
  tbl_visits,
  model,
  now = max(tbl_visits$t),
  eof_indicator = "EOF"
)

\method{compute_pfs}{model}(
  model,
  t,
  parameter_sample = NULL,
  warmup = 500L,
  nsim = 1000L,
  seed = NULL,
  ...
)

\method{plot_mstate}{model}(
  data,
  model,
  now = max(tbl_mstate$t_max),
  relative_to_sot = TRUE,
  ...
)

\method{plot_transition_times}{model}(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  confidence = NULL,
  ...
)

\method{plot_response_probability}{model}(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  ...
)

\method{plot_pfs}{model}(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  confidence = NULL,
  ...
)

\method{plot}{model}(
  x,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  confidence = NULL,
  ...
)
}
\arguments{
\item{x}{the model to plot}

\item{...}{further arguments passed to method implementations}

\item{model}{a model of class "Model"}

\item{n_per_group}{integer vector with number of individuals per group.}

\item{sample}{a stanfit object containing samples. These parameter samples
represent the parameter distribution over which the predictive distribution
averages. Technically, the parameters are resampled with replacement from
this sample to match the desired number of imputations.}

\item{nsim}{number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{nsim_parameters}{integer, number of parameter samples}

\item{warmup_parameters}{integer, number of warmup samples for the rstan sampler before retaining samples of the parameters.}

\item{as_mstate}{return data in multi-state forma, see \code{\link[=visits_to_mstate]{visits_to_mstate()}}}

\item{nuts_control}{control parameters for NUTS algorithm see \code{\link[rstan:stan]{rstan::stan()}}}

\item{data}{the (multi-state) data frame to impute further trajectories for.}

\item{now}{exact time point relative to start of the trial}

\item{recruitment_rates}{vector of recruitment rates}

\item{p_mean}{numeric, mean of the beta prior for the response probability}

\item{p_n}{numeric, beta prior equivalent sample size (a + b)}

\item{p_eta}{numeric, robustification parameter for beta prior; actual
prior is (1 - eta) beta + eta; i.e., eta is the non-informative weight.}

\item{p_min}{numeric, minimal response probability}

\item{p_max}{numeric, maximal response probability}

\item{median_t_q05}{numeric of length three,
5\% quantiles of the log-normal distributions for the
median time-to-next-event for the three transitions s->r, s->p, r->p.}

\item{median_t_q95}{numeric of length three,
95\% quantiles of the log-normal distributions for the
median time-to-next-event for the three transitions s->r, s->p, r->p.}

\item{shape_q05}{numeric of length three,
5\% quantiles of the log-normal distributions for the shapes of the
time-to-next-event distributions for the three transitions s->r, s->p, r->p.}

\item{shape_q95}{numeric of length three,
95\% quantiles of the log-normal distributions for the shapes of the
time-to-next-event distributions for the three transitions s->r, s->p, r->p.}

\item{visit_spacing}{numeric, fixed duration between visits}

\item{recruitment_rate}{numeric, constant recruitment rate}

\item{maximal_time}{the maximal overall runtime of the trial as measured from
the first visit of any group. No visits past this point are sampled.}

\item{tbl_visits}{visit data in long format}

\item{eof_indicator}{state name indicating (exactly observed) end of
follow up.}

\item{t}{a vector of time-points at which the PFS rate should be computed}

\item{parameter_sample}{a stanfit object with samples from the respective model.}

\item{warmup}{integer, number of warmup samples for the rstan sampler before retaining samples; these are used to
tune the hyperparameters of the MCMC algorithm.}

\item{relative_to_sot}{Boolean, should the timeline be relative to the start
of trial or the start of treatment for each individual}

\item{dt_interval}{numeric vector of length two with minimal and maximal
time (relative to individual first visit) to use for plotting}

\item{dt_n_grid}{number of grid points to use when automatically choosing
plotting interval}

\item{dt_expand}{expansion factor for upper plotting limit when using
automatic interval detection}

\item{dt_grid}{numeric vector of time points to use for plotting}

\item{confidence}{numeric in (0, 1) confidence level for point-wise
confidence bands around mean; none plotted if NULL.}
}
\value{
An object of type srp_group_prior holding all prior information in a
list-like structure; visit_spacing and recruitment_rate are accessible as
attributes.

an object of class srp_model; a named list of the group_ids,
the maximal time, the visit spacing, the recruitment rate, and a list of
prior parameters for the response probability, the median transition times,
and the shape of the transition time distribution (Weibull)
}
\description{
This function takes one or more prior-specifications for an SRP multi-state
model and combines them into a joint model.
Groups are still treated as independent.

This function takes one or more prior-specifications for an SRP multi-state
model and combines them into a joint model.
Groups are still treated as independent.
}
