% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\alias{group_prior}
\alias{create_model}
\alias{format.model}
\alias{print.model}
\title{A stable-response-progression model}
\usage{
group_prior(
  p_mean = 0.5,
  p_n = 3,
  p_eta = 0,
  p_min = 0,
  p_max = 1,
  median_t_q05 = c(1, 1, 3),
  median_t_q95 = c(12, 12, 24),
  shape_q05 = rep(0.99, 3),
  shape_q95 = rep(1.01, 3),
  visit_spacing = 1,
  recruitment_rate = 1
)

create_model(..., maximal_time = 10 * 12)

\method{format}{model}(x, ...)

\method{print}{model}(x, ...)
}
\arguments{
\item{p_mean}{numeric, mean of the beta prior for the response probability}

\item{p_n}{numeric, beta prior equivalent sample size (a + b)}

\item{p_eta}{numeric, robustification parameter for beta prior; actual
prior is (1 - eta) beta + eta; i.e., eta is the non-informative weight.}

\item{p_min}{numeric, minimal response probability}

\item{p_max}{numeric, maximal response probability}

\item{median_t_q05}{numeric of length three,
5\% quantiles of the log-normal distributions for the
median time-to-next-event for the three transitions s->r, s->p, r->p.}

\item{median_t_q95}{numeric of length three,
95\% quantiles of the log-normal distributions for the
median time-to-next-event for the three transitions s->r, s->p, r->p.}

\item{shape_q05}{numeric of length three,
5\% quantiles of the log-normal distributions for the shapes of the
time-to-next-event distributions for the three transitions s->r, s->p, r->p.}

\item{shape_q95}{numeric of length three,
95\% quantiles of the log-normal distributions for the shapes of the
time-to-next-event distributions for the three transitions s->r, s->p, r->p.}

\item{visit_spacing}{numeric, fixed duration between visits}

\item{recruitment_rate}{numeric, constant recruitment rate}

\item{...}{further arguments passed to method implementations}

\item{maximal_time}{the maximal overall runtime of the trial as measured from
the first visit of any group. No visits past this point are sampled.}

\item{x}{Model to print}
}
\description{
\code{create_model()} takes one or more prior-specifications for an
SRP multi-state model and combines them into a single model object.
Groups are still treated as independent.
}
\details{
\code{group_prior()} specifies a prior distribution for a
three state model (stable, response, progression) for
a single group.
}
\examples{
# a model with prior 25\% response rate and variance equivalent to
# 10 data points (i.e. a Beta(2.5, 7.5) distribution).
grp <- group_prior(p_mean = 0.25, p_n = 10)

# a model with two groups and different priors on the respective response
# probabilities
mdl <- create_model(
  A = group_prior(),
  B = group_prior(p_mean = 0.33, p_n = 10)
)

format(create_model(A = group_prior()))
print(create_model(A = group_prior()))
}
