% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model}
\alias{Model}
\alias{sample_prior.Model}
\alias{sample_posterior.Model}
\alias{sample_predictive.Model}
\alias{generate_visit_data.Model}
\title{An abstract multi-state model class}
\usage{
\method{sample_prior}{Model}(
  model,
  warmup = 500L,
  nsim = 2000L,
  seed = NULL,
  rstan_output = TRUE,
  pars = attr(model, "parameter_names"),
  ...
)

\method{sample_posterior}{Model}(
  model,
  data,
  warmup = 500L,
  nsim = 2000L,
  seed = NULL,
  rstan_output = TRUE,
  pars = attr(model, "parameter_names"),
  ...
)

\method{sample_predictive}{Model}(
  model,
  n_per_group,
  sample = NULL,
  nsim = 100L,
  nsim_parameters = 1000L,
  warmup_parameters = 250,
  seed = NULL,
  ...
)

\method{generate_visit_data}{Model}(model, n_per_group, seed = NULL, ...)
}
\description{
This is abstract class defining a standard set of methods for any implemented
multi-state model.
Objects of class 'Model' cannot be instantiated directly.
Only objects of the respective sub-classes can.
}
\seealso{
\link{srp_model}
}
