% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R, R/mstate_to_visit.R
\name{Model}
\alias{Model}
\alias{format.Model}
\alias{print.Model}
\alias{sample_prior.Model}
\alias{sample_posterior.Model}
\alias{sample_predictive.Model}
\alias{impute.Model}
\alias{parameter_sample_to_tibble.Model}
\alias{plot_mstate.Model}
\alias{generate_visit_data.Model}
\alias{sample_pfs_rate.Model}
\alias{visits_to_mstate.Model}
\alias{mstate_to_visits.Model}
\title{An abstract multi-state model class}
\usage{
\method{format}{Model}(x, ...)

\method{print}{Model}(x, ...)

\method{sample_prior}{Model}(
  model,
  warmup = 500L,
  nsim = 2000L,
  seed = NULL,
  rstan_output = TRUE,
  pars = attr(model, "parameter_names"),
  ...
)

\method{sample_posterior}{Model}(
  model,
  data,
  warmup = 500L,
  nsim = 2000L,
  seed = NULL,
  rstan_output = TRUE,
  pars = attr(model, "parameter_names"),
  ...
)

\method{sample_predictive}{Model}(
  model,
  n_per_group,
  sample = NULL,
  nsim = 100L,
  nsim_parameters = 1000L,
  warmup_parameters = 250,
  seed = NULL,
  as_mstate = FALSE,
  ...
)

\method{impute}{Model}(
  model,
  data,
  n_per_group = NULL,
  recruitment_rates = NULL,
  now = NULL,
  sample = NULL,
  nsim = 250L,
  nsim_parameters = 1000L,
  warmup_parameters = 250L,
  seed = NULL,
  ...
)

\method{parameter_sample_to_tibble}{Model}(model, sample, ...)

\method{plot_mstate}{Model}(data, model, now, relative_to_sot, ...)

\method{generate_visit_data}{Model}(model, n_per_group, recruitment_rate, seed = NULL, ...)

\method{sample_pfs_rate}{Model}(
  model,
  t,
  sample = NULL,
  warmup = 500L,
  nsim = 2000L,
  seed = NULL,
  ...
)

\method{visits_to_mstate}{Model}(
  tbl_visits,
  model,
  now = max(tbl_visits$t),
  eof_indicator = "EOF"
)

\method{mstate_to_visits}{Model}(model, tbl_mstate, ...)
}
\arguments{
\item{x}{Model to print}

\item{...}{further arguments passed to method implementations}

\item{model}{a model of class "Model"}

\item{warmup}{integer, number of warmup samples for the rstan sampler before retaining samples; these are used to
tune the hyperparameters of the MCMC algorithm.}

\item{nsim}{number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{rstan_output}{logical, function returns raw rstan sample object if set to TRUE}

\item{pars}{character vector of parameter names to sample; resorting to model defaults if NULL}

\item{data}{the (multi-state) data frame to impute further trajectories for.}

\item{n_per_group}{the number of individuals per group to be recruited.}

\item{sample}{a stanfit object containing samples. These parameter samples
represent the parameter distribution over which the predictive distribution
averages. Technically, the parameters are resampled with replacement from
this sample to match the desired number of imputations.}

\item{nsim_parameters}{integer, number of parameter samples}

\item{warmup_parameters}{integer, number of warmup samples for the rstan sampler before retaining samples of the parameters.}

\item{recruitment_rates}{the per-group recruitment rates.}

\item{now}{exact time point relative to start of the trial}

\item{relative_to_sot}{Boolean, should the timeline be relative to the start
of trial or the start of treatment for each individual}

\item{recruitment_rate}{numeric vector with the monthly recruitment rates
per group}

\item{t}{a vector of time-points at which the PFS rate should be computed}

\item{tbl_visits}{visit data in long format}

\item{eof_indicator}{state name indicating (exactly observed) end of
follow up.}

\item{tbl_mstate}{a data table in multi-state format. TODO: explain}
}
\description{
This is abstract class defining a standard set of methods for any implemented
multi-state model.
Objects of class 'Model' cannot be instantiated directly, only objects of
the respective sub-classes can be.
}
\seealso{
\link{srp_model}
}
