% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{plot_mstate}
\alias{plot_mstate}
\title{Swimmer-like plot of multi-state data}
\usage{
plot_mstate(data, model, now, relative_to_sot, ...)
}
\arguments{
\item{data}{a data table with multi-state data}

\item{model}{a model of class "Model"}

\item{now}{the current time relative to the start of the trial (sot)}

\item{relative_to_sot}{Boolean, should the timeline be relative to the start
of trial or the start of treatment for each individual}

\item{...}{further arguments passed to method implementations}
}
\description{
Swimmer-like plot of multi-state data
}
