% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{plot_pfs}
\alias{plot_pfs}
\title{Plot the progression-free-survial function of a model}
\usage{
plot_pfs(model, parameter_sample, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{parameter_sample}{a stanfit object with samples from the respective model.}

\item{...}{further arguments passed to method implementations}
}
\description{
Plot the progression-free-survial function of a model
}
