% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{plot_response_probability}
\alias{plot_response_probability}
\title{Plot the response probability distributions of a model}
\usage{
plot_response_probability(model, parameter_sample, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{parameter_sample}{a stanfit object with samples from the respective model.}

\item{...}{further arguments passed to method implementations}
}
\description{
Plot the response probability distributions of a model
}
