% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{plot_transition_times}
\alias{plot_transition_times}
\title{Plot the transition time distributions of a model}
\usage{
plot_transition_times(model, parameter_sample, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{parameter_sample}{a stanfit object with samples from the respective model.}

\item{...}{further arguments passed to method implementations}
}
\description{
Plot the transition time distributions of a model
}
