% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample model posterior parameters}
\usage{
sample_posterior(model, data, nsim, seed, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{data}{a data frame with variables "subject_id", "group_id", "t_recruitment", "dt1" and "dt2"
where dt1 is the minimal and dt2 the maximal time to the event in question.
If both are Inf, the individual has not and will never experience the event.
If dt1 < Inf but dt2 == Inf the individual is still at risk.
"t_recruitment", "dt1" and "dt2" can also be missing if individual is not yet recruited.}

\item{nsim}{number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}
}
\value{
A rstanfit object with posterior samples.
}
\description{
Posterior sample model of the parameters conditional on a data set.
}
\seealso{
\code{\link[=parameter_sample_to_tibble]{parameter_sample_to_tibble()}}
}
