% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample model posterior parameters}
\usage{
sample_posterior(model, data, warmup, nsim, seed, rstan_output, pars, ...)
}
\arguments{
\item{model}{the model to sample from}

\item{data}{data frame with columns "group_id", "subject_id",
"t_recruitment" (recruitment time since start of trial),
"dt1" (minimal time since recruitment to event, Inf if definite no event),
"dt2" (maximal time since recruitment to event, Inf if definite no event or censored)}

\item{nsim}{number of samples to draw}

\item{rstan_output}{return raw rstan output?}
}
\value{
...
}
\description{
Sample model parameters conditional on a data set (posterior).
}
