% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{sample_predictive}
\alias{sample_predictive}
\title{Sample data from predictive distribution of a model}
\usage{
sample_predictive(
  model,
  n_per_group,
  sample,
  nsim,
  nsim_parameters,
  warmup_parameters,
  seed,
  as_mstate,
  ...
)
}
\arguments{
\item{model}{a model of class "Model"}

\item{n_per_group}{integer vector with number of individuals per group.}

\item{sample}{a stanfit object with samples from the respective model.}

\item{nsim}{number of samples to draw}

\item{nsim_parameters}{integer, number of parameter samples}

\item{warmup_parameters}{integer, number of warmup samples for the rstan sampler before retaining samples of the parameters.}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{as_mstate}{return data in multi-state forma, see \code{\link[=visits_to_mstate]{visits_to_mstate()}}}

\item{...}{further arguments passed to method implementations}
}
\value{
TODO:
}
\description{
Sample data from predictive distribution of a model
}
