% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_prior.R
\name{sample_prior}
\alias{sample_prior}
\title{Sample from prior distribution of a model}
\usage{
sample_prior(
  model,
  nsim = 2000L,
  seed = NULL,
  warmup = 500L,
  pars = attr(model, "parameter_names"),
  nuts_control = list(),
  ...
)
}
\arguments{
\item{model}{an object of class \link{srpmodel} containing prior information}

\item{nsim}{integer, number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{warmup}{integer, number of warm-up samples for the MCMC sampler
before retaining samples; see \code{warmup} parameter in \code{\link[rstan:stan]{rstan::stan()}}.}

\item{pars}{character vector of parameter names to sample}

\item{nuts_control}{list, parameters for NUTS algorithm see \code{control}
argument in\code{\link[rstan:stan]{rstan::stan()}}}

\item{...}{further arguments passed to method implementations}
}
\value{
A \link[rstan:stanfit-class]{rstan::stanfit} object with sampled prior parameters.
}
\description{
\code{sample_prior()} draws samples from the
prior distribution of the specified model object.
}
\examples{
mdl <- create_srpmodel(A = define_srp_prior())
sample_prior(mdl, 1000L, seed = 42L)

}
\seealso{
\code{\link[rstan:stan]{rstan::stan()}} \code{\link[=parameter_sample_to_tibble]{parameter_sample_to_tibble()}} \code{\link[=sample_posterior]{sample_posterior()}}
\code{\link[=sample_predictive]{sample_predictive()}} \code{\link[=impute]{impute()}}
}
