% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{sample_prior}
\alias{sample_prior}
\title{Sample model prior parameters}
\usage{
sample_prior(model, nsim, seed, ...)
}
\arguments{
\item{model}{a model of class "Model"}

\item{nsim}{number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}
}
\value{
a rstanfit object with the sampled prior parameters
}
\description{
Sample model parameters from the model prior distribution.
}
\seealso{
\code{\link[=parameter_sample_to_tibble]{parameter_sample_to_tibble()}}
}
