% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_prior.R
\name{sample_prior}
\alias{sample_prior}
\alias{sample_prior.model}
\title{Sample from prior distribution of a model}
\usage{
sample_prior(model, nsim, seed, ...)

\method{sample_prior}{model}(
  model,
  nsim = 2000L,
  seed = NULL,
  warmup = 500L,
  pars = attr(model, "parameter_names"),
  nuts_control = list(),
  ...
)
}
\arguments{
\item{model}{a model of class "Model"}

\item{nsim}{number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{...}{further arguments passed to method implementations}

\item{warmup}{integer, number of warmup samples for the rstan sampler before retaining samples; these are used to
tune the hyperparameters of the MCMC algorithm.}

\item{pars}{character vector of parameter names to sample; resorting to model defaults if NULL}

\item{nuts_control}{control parameters for NUTS algorithm see \code{\link[rstan:stan]{rstan::stan()}}}
}
\value{
A rstanfit object with sampled prior parameters.
}
\description{
\code{sample_prior()} draws samples from the
prior distribution of the specified model.
}
\examples{
mdl <- create_srp_model(A = srp_group_prior())
sample_prior(mdl, 1000L, seed = 42L)

}
\seealso{
\code{\link[=parameter_sample_to_tibble]{parameter_sample_to_tibble()}}
}
