% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{sample_prior_predictive}
\alias{sample_prior_predictive}
\title{Sample model prior parameters}
\usage{
sample_prior_predictive(model, n_per_arm, ...)
}
\arguments{
\item{model}{the model to sample from}

\item{nsim}{number of samples to draw}

\item{rstan_output}{return raw rstan output?}
}
\value{
...
}
\description{
Sample model parameters unconditionally (prior).
}
