% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_model.R, R/srp_model_plotting.R
\name{srp_model}
\alias{srp_model}
\alias{create_srp_model}
\alias{format.srp_model}
\alias{visits_to_mstate.srp_model}
\alias{compute_pfs.srp_model}
\alias{plot_mstate.srp_model}
\alias{plot_transition_times.srp_model}
\alias{plot_response_probability.srp_model}
\alias{plot_pfs.srp_model}
\alias{plot.srp_model}
\title{A Stable-Response-Progression Model}
\usage{
create_srp_model(
  group_id,
  logodds_mean,
  logodds_sd = rep(0.01, length(group_id)),
  median_time_to_next_event_mean,
  median_time_to_next_event_sd = matrix(0.1, nrow = length(group_id), ncol = 3),
  visit_spacing,
  recruitment_rate = rep(1, length(group_id)),
  max_time = 10 * 12,
  logodds_min = rep(logodds(0.001), length(group_id)),
  logodds_max = rep(logodds(0.999), length(group_id)),
  shape_min = matrix(0.99, nrow = length(group_id), ncol = 3),
  shape_max = matrix(1.01, nrow = length(group_id), ncol = 3)
)

\method{format}{srp_model}(x, ...)

\method{visits_to_mstate}{srp_model}(
  tbl_visits,
  model,
  now = max(tbl_visits$t),
  eof_indicator = "EOF"
)

\method{compute_pfs}{srp_model}(
  model,
  t,
  parameter_sample = NULL,
  warmup = 500L,
  nsim = 1000L,
  seed = NULL,
  ...
)

\method{plot_mstate}{srp_model}(
  data,
  model,
  now = max(tbl_mstate$t_max),
  relative_to_sot = TRUE,
  ...
)

\method{plot_transition_times}{srp_model}(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  ...
)

\method{plot_response_probability}{srp_model}(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  ...
)

\method{plot_pfs}{srp_model}(
  model,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  ...
)

\method{plot}{srp_model}(
  x,
  parameter_sample = NULL,
  seed = 42L,
  nsim = 500L,
  warmup = 250,
  nuts_control = list(),
  dt_interval = NULL,
  dt_n_grid = 25,
  dt_expand = 1.1,
  dt_grid = NULL,
  ...
)
}
\arguments{
\item{group_id}{a character vector with the group ids, these are used to
check compatibility of data later}

\item{logodds_mean}{a vector with the means of the (truncated) normal priors
on the log-odds of the response probability}

\item{logodds_sd}{a vector with the standard deviations of the (truncated)
normal priors on the log-odds of the response probability}

\item{median_time_to_next_event_mean}{a matrix with the means of the
(truncated) normal priors on the median time to next event for each of the
Weibull transition probabilities, the (i,j)-th entry is the i-th group
median time to next event for transition j
(1=stable-response, 2=stable-progression, 3=response-progression)}

\item{median_time_to_next_event_sd}{a matrix with the standard deviations of
the (truncated) normal priors on the median time to next event for each of
the Weibull transition probabilities, the (i,j)-th entry is the i-th group
median time to next event for transition j
(1=stable-response, 2=stable-progression, 3=response-progression)}

\item{visit_spacing}{vector of time differences between visits per group,
only relevant for sampling from the predictive distribution}

\item{recruitment_rate}{vector with per-group recruitment rates,
only relevant for sampling from the predictive distribution}

\item{max_time}{maximal overall runtime from first visit,
only relevant for sampling from the predictive distribution}

\item{logodds_min}{lower boundary on the log-odds per group}

\item{logodds_max}{upper boundary on the log-odds per group}

\item{shape_min}{matrix of lower boundaries of the uniform prior of the
Weibull distribution per group/transition}

\item{shape_max}{matrix of upper boundaries of the uniform prior of the
Weibull distribution per group/transition}

\item{x}{the model to plot}

\item{...}{further arguments passed to method implementations}

\item{tbl_visits}{visit data in long format}

\item{model}{a multi-state model object}

\item{now}{time point since start of trial (might be later than last
recorded visit)}

\item{eof_indicator}{state name indicating (exactly observed) end of
follow up.}

\item{t}{a vector of time-points at which the PFS rate should be computed}

\item{parameter_sample}{a stanfit object with samples from the respective model.}

\item{warmup}{integer, number of warmup samples for the rstan sampler before retaining samples; these are used to
tune the hyperparameters of the MCMC algorithm.}

\item{nsim}{number of samples to draw}

\item{seed}{integer, fixed random seed; NULL for no fixed seed}

\item{data}{a data table with multi-state data}

\item{relative_to_sot}{Boolean, should the timeline be relative to the start
of trial or the start of treatment for each individual}

\item{nuts_control}{control parameters for NUTS algorithm see \code{\link[rstan:stan]{rstan::stan()}}}

\item{dt_interval}{numeric vector of length two with minimal and maximal
time (relative to individual first visit) to use for plotting}

\item{dt_n_grid}{number of grid points to use when automatically choosing
plotting interval}

\item{dt_expand}{expansion factor for upper plotting limit when using
automatic interval detection}

\item{dt_grid}{numeric vector of time points to use for plotting}
}
\description{
Create a new instance of an SRP model
}
\details{
TODO
}
\seealso{
\link{Model}
}
