% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstate_to_visit.R, R/srp_model.R
\name{mstate_to_visits.srp_model}
\alias{mstate_to_visits.srp_model}
\alias{srp_model}
\alias{create_srp_model}
\alias{format.srp_model}
\alias{plot_mstate.srp_model}
\alias{visits_to_mstate.srp_model}
\title{A Stable-Response-Progression Model}
\usage{
\method{mstate_to_visits}{srp_model}(model, tbl_mstate, ...)

create_srp_model(
  group_id,
  logodds_mean,
  logodds_sd = rep(0.01, length(group_id)),
  median_time_to_next_event_mean,
  median_time_to_next_event_sd = matrix(0.01, nrow = length(group_id), ncol = 3),
  visit_spacing,
  recruitment_rate = rep(1, length(group_id)),
  max_time = 10 * 12,
  logodds_min = rep(logodds(0.001), length(group_id)),
  logodds_max = rep(logodds(0.999), length(group_id)),
  shape_min = matrix(0.99, nrow = length(group_id), ncol = 3),
  shape_max = matrix(1.01, nrow = length(group_id), ncol = 3)
)

\method{format}{srp_model}(x, ...)

\method{plot_mstate}{srp_model}(
  data,
  model,
  now = max(tbl_mstate$t_max),
  relative_to_sot = TRUE,
  ...
)

\method{visits_to_mstate}{srp_model}(
  tbl_visits,
  model,
  now = max(tbl_visits$t),
  eof_indicator = "EOF"
)
}
\arguments{
\item{model}{a model of class "Model"}

\item{tbl_mstate}{a data table in multi-state format. TODO: explain}

\item{...}{further arguments passed to method implementations}

\item{group_id}{a character vector with the group ids, these are used to
check compatibility of data later}

\item{logodds_mean}{a vector with the means of the (truncated) normal priors
on the log-odds of the response probability}

\item{logodds_sd}{a vector with the standard deviations of the (truncated)
normal priors on the log-odds of the response probability}

\item{median_time_to_next_event_mean}{a matrix with the means of the
(truncated) normal priors on the median time to next event for each of the
Weibull transition probabilities, the (i,j)-th entry is the i-th group
median time to next event for transition j
(1=stable-response, 2=stable-progression, 3=response-progression)}

\item{median_time_to_next_event_sd}{a matrix with the standard deviations of
the (truncated) normal priors on the median time to next event for each of
the Weibull transition probabilities, the (i,j)-th entry is the i-th group
median time to next event for transition j
(1=stable-response, 2=stable-progression, 3=response-progression)}

\item{visit_spacing}{vector of time differences between visits per group}

\item{logodds_min}{lower boundary on the log-odds per group}

\item{logodds_max}{upper boundary on the log-odds per group}

\item{shape_min}{matrix of lower boundaries of the uniform prior of the
Weibull distribution per group/transition}

\item{shape_max}{matrix of upper boundaries of the uniform prior of the
Weibull distribution per group/transition}

\item{x}{SRP model to format}

\item{data}{a data table with multi-state data}

\item{now}{the current time relative to the start of the trial (sot)}

\item{relative_to_sot}{Boolean, should the timeline be relative to the start
of trial or the start of treatment for each individual}

\item{tbl_visits}{visit data in long format}

\item{eof_indicator}{state name indicating (exactly observed) end of
follow up.}
}
\description{
Create a new instance of an SRP model
}
\details{
TODO
}
\seealso{
\link{Model}
}
