% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{visits_to_mstate}
\alias{visits_to_mstate}
\title{Convert cross sectional visit data to time-to-event data}
\usage{
visits_to_mstate(
  tbl_visits,
  model,
  now = max(tbl_visits$t),
  eof_indicator = "EOF"
)
}
\arguments{
\item{tbl_visits}{visit data in long format}

\item{model}{a multi-state model object}

\item{now}{time point since start of trial (might be later than last
recorded visit)}

\item{eof_indicator}{state name indicating (exactly observed) end of
follow up.}
}
\value{
A data frame
}
\description{
This function assumes that the visit density is high enough to not miss any
transient state jumps.
}
