% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits_to_mstate.R
\name{visits_to_mstate}
\alias{visits_to_mstate}
\title{Convert longitudinal visit data to time-to-event data}
\usage{
visits_to_mstate(
  tbl_visits,
  start_state,
  absorbing_states,
  now = max(tbl_visits$t),
  eof_indicator = "EOF"
)
}
\arguments{
\item{tbl_visits}{visit data in long format}

\item{start_state}{staring state}

\item{absorbing_states}{character vector of absorbing states}

\item{now}{time point since start of trial (might be later than last
recorded visit)}

\item{eof_indicator}{state name indicating (exactly observed) eond of
follow up.}
}
\value{
A data frame
}
\description{
Convert longitudinal visit data to time-to-event data
}
